+++ apps/client/src/environments/patch-notes.ts
@@ -1,12 +1,8 @@
-export const patchNotes = `### Bug Fixes
+export const patchNotes = `### Features
 
-* **collectables:** fix alarm buttons not showing on timed items.
-* **db:** fix Islekeep's Stone Hatchet recipe.
-* **island-workshop:** display error message when no plannings are found.
-* **log-tracker:** fixed selected recipes not resetting on page change.
+* **island-workshop:** optimizer configuration & island level filter.
 
 
-### Features
+### Performance Improvements
 
-* **allagan-reports:** support for island pasture animal source.
-* **lists:** add support for landmark recipes.`;
+* **core:** removed recent changes that made the app slower.`;